+++ src/HasteModuleLoader/__tests__/HasteModuleLoader-NODE_PATH-test.js
@@ -10,14 +10,14 @@
 'use strict';
 
 jest.autoMockOff();
-jest.mock('../../JSDomEnvironment');
+jest.mock('../../environments/JSDOMEnvironment');
 
 var path = require('path');
 var utils = require('../../lib/utils');
 
 describe('HasteModuleLoader', function() {
   var HasteModuleLoader;
-  var JSDomEnvironment;
+  var JSDOMEnvironment;
   var resourceMap;
 
   var CONFIG = utils.normalizeConfig({
@@ -33,7 +33,7 @@ describe('HasteModuleLoader', function() {
         return buildLoader();
       });
     } else {
-      var mockEnvironment = new JSDomEnvironment(CONFIG);
+      var mockEnvironment = new JSDOMEnvironment(CONFIG);
       return Promise.resolve(
         new HasteModuleLoader(CONFIG, mockEnvironment, resourceMap)
       );
@@ -43,7 +43,7 @@ describe('HasteModuleLoader', function() {
   function initHasteModuleLoader(nodePath) {
     process.env.NODE_PATH = nodePath;
     HasteModuleLoader = require('../HasteModuleLoader');
-    JSDomEnvironment = require('../../JSDomEnvironment');
+    JSDOMEnvironment = require('../../environments/JSDOMEnvironment');
   }
 
   pit('uses NODE_PATH to find modules', function() {