+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.7.1
+
+* Correctly map `process.on` into jsdom environments, fixes a bug introduced in
+  jest 0.7.0.
+
 ## 0.7.0
 
 * Fixed a memory leak with test contexts. Jest now properly cleans up